+++ bundles/framework/myplaces3/view/MainView.js
@@ -166,7 +166,7 @@ Oskari.clazz.define('Oskari.mapframework.bundle.myplaces3.view.MainView',
             const sandbox = this.instance.sandbox;
             sandbox.postRequestByName('DisableMapKeyboardMovementRequest');
             this.form = Oskari.clazz.create(
-                'Oskari.mapframework.bundle.myplaces3.view.PlaceFormV2',
+                'Oskari.mapframework.bundle.myplaces3.view.PlaceForm',
                 this.instance,
                 this.options
             );
